requirements.txt
elasticsearch==8.13.0
fastapi==0.111.1
starlette==0.37.2
pandas==2.2.2
mysql-connector-python==9.0.0
mysqlclient==2.2.4
numpy==1.26.4
scipy==1.13.0
scikit-learn==1.5.1
nltk==3.8.1